remove circular dependencies in util
